LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-single-value-in-body-expected.txt


FAIL Meta color-scheme in body should apply. assert_equals: expected "dark" but got "auto"
